Backend/Backend/credential/serializers.py
from rest_framework import serializers
from .models import Certificate
from .models import Badge

class CertificateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Certificate
        fields = '__all__'


class BadgeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Badge
        fields = '__all__'