Backend/Backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
python-decouple==3.8
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
